Compact JSON-LD signed incoming activities
